fix: remove bugs/annoyances from clean step

+++ clean.bat
@@ -4,7 +4,7 @@
 :: script on windows, because this is achievable from within the zig build system 
 :: on other platforms.
 
-rmdir /s /q .zig-cache
-rmdir /s /q .zig-out
-rmdir /s /q zig-cache
-rmdir /s /q zig-out
+if exist .zig-cache    rmdir /s /q .zig-cache
+if exist .zig-out      rmdir /s /q .zig-out
+if exist zig-cache     rmdir /s /q zig-cache
+if exist zig-out       rmdir /s /q zig-out
